mffc_utils.
