+++ include/mockturtle/algorithms/cut_rewriting.hpp
@@ -46,6 +46,7 @@
 #include "../utils/stopwatch.hpp"
 #include "../views/cut_view.hpp"
 #include "cut_enumeration.hpp"
+#include "mffc_utils.hpp"
 
 #include <fmt/format.h>
 
@@ -97,76 +98,6 @@ struct cut_rewriting_stats
 
 namespace detail
 {
-
-template<typename Ntk, typename TermCond>
-uint32_t recursive_deref( Ntk const& ntk, node<Ntk> const& n, TermCond&& terminate )
-{
-  /* terminate? */
-  if ( terminate( n ) )
-    return 0;
-
-  /* recursively collect nodes */
-  uint32_t value{1};
-  ntk.foreach_fanin( n, [&]( auto const& s ) {
-    if ( ntk.decr_value( ntk.get_node( s ) ) == 0 )
-    {
-      value += recursive_deref( ntk, ntk.get_node( s ), terminate );
-    }
-  } );
-  return value;
-}
-
-template<typename Ntk, typename TermCond>
-uint32_t recursive_ref( Ntk const& ntk, node<Ntk> const& n, TermCond&& terminate )
-{
-  /* terminate? */
-  if ( terminate( n ) )
-    return 0;
-
-  /* recursively collect nodes */
-  uint32_t value{1};
-  ntk.foreach_fanin( n, [&]( auto const& s ) {
-    if ( ntk.incr_value( ntk.get_node( s ) ) == 0 )
-    {
-      value += recursive_ref( ntk, ntk.get_node( s ), terminate );
-    }
-  } );
-  return value;
-}
-
-template<typename Ntk, typename LeavesIterator>
-uint32_t recursive_deref( Ntk const& ntk, node<Ntk> const& n, LeavesIterator begin, LeavesIterator end )
-{
-  return recursive_deref( ntk, n, [&]( auto const& n ) { return std::find( begin, end, n ) != end; } );
-}
-
-template<typename Ntk, typename LeavesIterator>
-uint32_t recursive_ref( Ntk const& ntk, node<Ntk> const& n, LeavesIterator begin, LeavesIterator end )
-{
-  return recursive_ref( ntk, n, [&]( auto const& n ) { return std::find( begin, end, n ) != end; } );
-}
-
-template<typename Ntk>
-uint32_t recursive_deref( Ntk const& ntk, node<Ntk> const& n )
-{
-  return recursive_deref( ntk, n, [&]( auto const& n ) { return ntk.is_constant( n ) || ntk.is_pi( n ); } );
-}
-
-template<typename Ntk>
-uint32_t recursive_ref( Ntk const& ntk, node<Ntk> const& n )
-{
-  return recursive_ref( ntk, n, [&]( auto const& n ) { return ntk.is_constant( n ) || ntk.is_pi( n ); } );
-}
-
-template<typename Ntk>
-uint32_t mffc_size( Ntk const& ntk, node<Ntk> const& n )
-{
-  auto v1 = recursive_deref( ntk, n );
-  auto v2 = recursive_ref( ntk, n );
-  assert( v1 == v2 );
-  (void)v2;
-  return v1;
-}
 
 class graph
 {
